// src/types/enum.ts
export enum EOperacaoCheque {
  A_PAGAR = "A_PAGAR",
}

export enum EStatusCheque {
  A_VENCER = "A_VENCER",
  VENCIDO = "VENCIDO",
  PAGO = "PAGO",
  DEVOLVIDO = "DEVOLVIDO",
}

export enum EPermissaoAcesso {
  GERENCIAR_ACESSOS = "GERENCIAR_ACESSOS",
  GERENCIAR_PERMISSOES = "GERENCIAR_PERMISSOES",
  GERENCIAR_RESPONSAVEIS = "GERENCIAR_RESPONSAVEIS",
  GERENCIAR_BANCOS = "GERENCIAR_BANCOS",
  GERENCIAR_CONTAS_BANCARIAS = "GERENCIAR_CONTAS_BANCARIAS",
  GERENCIAR_CHEQUES = "GERENCIAR_CHEQUES",
  GERENCIAR_DATAS_BLOQUEADAS = "GERENCIAR_DATAS_BLOQUEADAS",
  VISUALIZACAO_TOTAL = "VISUALIZACAO_TOTAL",
}